forpython/for-pdf-viewer/requirements.txt
pdf2image
pillow
click
tqdm
